big changes ...
git-svn-id: https://brezn.muc.ccc.de/svn/uberbus/trunk@93 0ae26b70-98e1-47eb-8c5c-8e5c6e8157e6

+++ firmware2/lib/serial_handler.h
@@ -21,7 +21,7 @@ inline void serial_putStop(void);
 void serial_sendFrames(char * s);
 void serial_sendFramec(uint8_t s);
 void serial_putenc(uint8_t * d, uint16_t n);
-unsigned int readline( void );
+uint16_t readline( void );
 uint16_t serial_readline(void);
 
 #endif
